[mojo-lang] Remove "'var' was never mutated, switch to a 'let'" warning. (#29278)
We are planning to remove `let` declarations entirely, so this is a tiny
step towards that.  Beyond that, this warning was never sound in the
presence of closures, is a nuisance that is causing some build warnings.
Just drop it.

MODULAR_ORIG_COMMIT_REV_ID: c464639f5a7e56089c0df67902aceaef7a17dae4