[mojo-lang][Checklifetimes][VarArgs] Have refitem extend lifetime of vararg (#29297)
In the case of owned variadics, the destructor of the VariadicListMem
is what destroys all of the captured value in the vararg pack, which
means that destructor can only be run after all the references to
the elements are done.  So far, we have had no way to express this,
and thus the VariadicListMem (and thus the elements in an owned
vararg) get destroyed too early.

This patch is a step forward (but still not quite enough) to solve this:
When returning a reference to an item in the set with refitem, it mixes
in the lifetime of the `VariadicListMem` itself into the reference.
CheckLifetimes isn't quite smart enough to do the right thing about this
but this should in principle allow it to know that a use of this exposed
reference keeps the Mem alive.

Getting this in place requires several esoteric things:
1) We need to form lifetime unions, so this adds a "parametric alias"
   as a convenience helper do compute this.
2) This needs to get the self reference in refitem, so it switches it
   to our janky syntax for this.
3) That exposes that LifetimeTracking isn't handling that right, so fix
   that to enable janky but effective stop-gaps.

This is enough to get the del for the mem pack after the iadd in
`variadic_inout_mems` (which is pretty huge) but not enough to handle
the iter case in `variadic_inout_mems_iter`.  Further work will address
this.

MODULAR_ORIG_COMMIT_REV_ID: 955f4ca60c99a8e9a34f26e5e37f6319ac450c66