[KGEN] Fold away `kgen.variant.is` proceeded by `kgen.variant.create` (#29403)
The existing folder of `VariantIsOp` only works for constants. This
patch improves it to work for non-constant but obvious case -- when the
input to `VariantIsOp` is a `VariantCreateOp` which contains the variant
index.

MODULAR_ORIG_COMMIT_REV_ID: d4e1a9ca4ed8b207130bd4962884f8e8d6d9312c