[mojo-lang] improve error message when materializing IntLiteral (#28622)
This catches materialization of `!kgen.int_literal` during lowering to LLVM, adding a clearer error message.  This way of catching it is somewhat disappointing, since we will need to add similar logic for each nonmaterializable type.  But catching nonmaterializable types earlier in the pipeline is difficult because we want to catch only instances that would be materialized, not interpreted at compile time, and there is no other centralized place to do that.

Fixes [Internal Link]

modular-orig-commit: 1c1216d3885d8e8ef92ac9802d1cb28a64c6e827