[KGEN][DebugInfo] Emit SourceName for structs (#26534)
Fixes #25054.

Emit struct names in DI using SourceName encoding.

Since we don't have struct scoping in DI, the proposed change attaches
the SourceName of the struct onto the LIT struct.decl op during parsing.
This should not incur much memory or runtime cost, since the SourceName
of structs are already calculated as part of calculating the SourceName
for struct methods. This change merely saves it onto the struct decl.

During LowerLITTypes, DeclRefTypes are converted into DIStructTypes. At
this point, the SourceName associated with the struct decl is augmented
with the parameter values, and attached onto the DIStructType. If the
decl did not come with a SourceName, that means the decl was parsed
without debuginfo. We fallback to just using the symbol name itself.

MODULAR_ORIG_COMMIT_REV_ID: 89ccd44095ef788acf908cda2e84ad79d498bed3