[Kernels] Re-enable mandelbrot notebook (#29487)
This notebook test was disabled in #26469 due to intermittent failures,
described as `numpy circular dependency` and:

```
benchmark_matmul_python() missing 3 required positional arguments: 'M', 'N', and 'K'`
```

In the time since this notebook was disabled, the Mojo source code has
bitrot a bit. This patch brings it up-to-date such that it compiles, and
also re-enables the test -- either the intermittent failures have been
fixed, or we'll take the time to investigate & fix once we get a new CI
failure.

Knocks a notebook off the list in #27873.

modular-orig-commit: 13ef6530f060819602f8f5f4191fae64c88806a3