[KGEN] Add test for `mojo package -o` (#29510)
This test is inspired by a question I received: does the original
directory name fed as an input to `mojo package` matter, or is the
output file name specified by `-o` the only thing that matters? The
documentation for the `mojo package -o` option states that the output
file name determines the name by which the package can be imported --
but as far as I could tell there isn't a test for this behavior, so this
patch adds one.

MODULAR_ORIG_COMMIT_REV_ID: a5842ec60c3b607f83be324c517f83e7f72cb6ff