[Mojo] Modify how alias names are mangled in the parser (#29536)
This patch starts to phase in a new parameter mangling scheme in the
parser. The new scheme will use a postpended backtick, along with the
scope depth and a unique counter to ensure parameter names don't shadow
and to improve hash stability. This patch starts to phase this in by
modifying how alias names are mangled. Currently, parameter names are
still non-lexical, so it is not possible to reliably check if a
shadowing name is declared by an alias in a parent scope. Because of
this, aliases are currently _always_ mangled, but this restriction will
be lifted when non-lexical parameters become prohibited.

MODULAR_ORIG_COMMIT_REV_ID: 85f421d7ec14068542c23e87ad9abde731f7baa7