[Mojo] Slightly improve the interface of `PassingKindPrinter` (#29149)
It was a bit goofy that the `printOptionalStarSlash` method needed both
the passing kind and an index. There was only a single use case that
necessitated this, which is now modified, so that `PassingKindPrinter`
now takes the array of passing kinds.

MODULAR_ORIG_COMMIT_REV_ID: a9e048d38d9c16d6d958f3e9ad0908134985dff6