[KGEN] Simplify COND via "True" forwarding in Then Branch (#29632)
The main idea behind this PR is if we have a KGEN expr `TEST: i1, THEN, ELSE`

Then in `cond(TEST, THEN, ELSE)` we can use the fact that `TEST == 1` in `THEN`.

This PR implements this idea for conditions where `TEST` is a measure of equality between two indices, one of which is a parameter and the other is a constant integer. For example `eq(A, 5)`
We then use this fact to substitute `A` with `5` recursively in `THEN` up to a fixed recursion limit.

This is useful for shape inference in many scenarios.
For example the broadcasting rules for dimensions are essentially:

```
def broadcast(D0: index, D1: index):
    # add some assertions
    if D0 == 1:
       return D1
    return D0
```

As `D1` may be already parameterized in terms of `D0` this is useful for simplification.

MODULAR_ORIG_COMMIT_REV_ID: 6489e5dc53ba3a2668c51bdd39702ced1806d5b1