[VSCode] Fix import of the ps-list package (#29593)
The ps-list package worked for me last week when I trying the extension
in debug mode, however, today I discovered that it doesn't work as a
regular vsix installation.
The reason is that ps-list is a modern ES module and not a commonjs
module, which is what Electron needs from a vsix.
The only two workarounds at this point are copy/pasting ps-list into our
tree or creating a fork of ps-list and turn it into a commonjs module.
The former is way easier, so I'm doing that.

I also removed a user-visible error when the process picker was
cancelled, and cleaned up some minor related things.

MODULAR_ORIG_COMMIT_REV_ID: da65bc42f12f2b1a5bc0b701ad6358dc9f8c946c