[Telemetry] Remove body, severity, and streaming from emitEvent API (#29610)
For emitting events, we should not use the log body field. Instead, we
should have structured fields that contain the data we wish to include.
For emitting events, we don't use the logging "severity" field. Instead
the information that this is "an error" should be in the name of the
log. Migrate all error messages to a structured "error" field.
The streaming API encourages use of the body field of logs. Migrate the
1 usage to emitEvent & remove the LogStream API

MODULAR_ORIG_COMMIT_REV_ID: 076e63a74340a2fc37de52daaeedd22faacbad48