[Support][LLCL] Move ConcurrentAppendingVector to Support/ADT. (#29735)
This is a generic data structure that we use in various places - it belongs to the set of 'general goodness' that usually lives in Support.

Besides the high-brow reasons, we're trying to move TypeID to Support so we can move GenericUniquePtr over - and it needs this.

MODULAR_ORIG_COMMIT_REV_ID: 288001bce20a36c0c6102ab6bb8483db349e0fff