Revert "[KGEN][mojo-lang] make pointer in struct opaque" (#29750)
I'm reverting this because this breaks debug info.

Specifically:

If you define a `DynamicVector[Int]`, before this PR, in DWARF, we
produced a `data` member child for this struct type that contains the
type `(index *)`, which is what we need to be able to display the items
of the vector when debugging. After this PR, we produce `void *`, so
there's no way we can display the items of the vector.

A proper fix for this PR would be to leave DWARF unchanged.

MODULAR_ORIG_COMMIT_REV_ID: 4785bd1158e98a3d085a1623bf8603a893ba6117