[Mojo] Escape parameter names when printing metatype decl refs (#29811)
In particular, this is important because we use `type` as sugar for
`kgen.type`, and therefore parameters named `type` should be escaped
(i.e. `*"type"`) when used in expressions that are metatypes. This patch
introduces a new `isMetaType` method on the `ParameterTypeInterface`,
and updates the parameter name printing to use it.

MODULAR_ORIG_COMMIT_REV_ID: 1285d929dcaa4d9f7028eda8961bd1027551c0db