[Docs] Dict API doc edits (#29717)
This is a first pass at making the Dict API docs a little more
user-friendly. Added a few basics and moved the implementation details
from the public docstring to an internal comment. I also made a minor
tweak to the module docstring so that only one sentence shows up in the
summary (instead of the entire first paragraph.) I didn't look at all
the methods, etc. This is a quick PR to try and make Dict easier to use
in 0.7.0.

Public doc string versus code comment: I think the exact division here
is open to debate, but as written the API docs included a lot of
references to private variables inside Dict, which seems confusing for
people who don't have access to the source. Some of this could be
rewritten in other terms, but it may have more value to maintainers if
it shows the actual variable names.

Long-term, we probably need to establish some API doc style guidelines
here.

I left the information about maintaining insertion order out of the
public docstring because, until we have an iterator, it's not meaningful
to the end user.

modular-orig-commit: 88c3407bc5e7d21e2f3f48cafe14f52ce2167025