[Support] Revamp the EntitlementStore API. (#29809)
This patch updates the EntitlementStore API a little bit - as we use it in more cases we are discovering that the current API doesn't quite fit. Instead of massaging the users, massage the thing we're using so it composes a little bit better.

To summarize the change - rather than having `open` implicitly attempt to fetch a new cert, have `open` open an existing cert and add a separate API `generate` which will generate a new cert. This allows clients to choose their own policy on how to behave if we don't have an existing cert; this is important in this transition period we're in right now where users haven't fully switched over.

MODULAR_ORIG_COMMIT_REV_ID: 6641a0cc4a219dd51c65ba881ebabe04d6abf23a