[KGEN][DebugInfo] Mem2reg DIExpr LeafReplacer should not force initial undef value (#29877)
**Background**: Retiring `let` and moving everything into a `var` has a
non-zero impact on debuginfo because mem2reg does not restore the
previous IR identically (which meant it didn't trigger the debug value
to debug declare transformation when lowering to llvm):

Previously, we generated debuginfo ValueOp for `let` decls by directly
referring to the value it is being initialized with. This means there's
only ever one DebugInfo::ValueOp for each let decl.

With `var` decls, we start with one DebugInfo::ValueOp pointing to the
stack allocation of this variable. But mem2reg will always leave us with
N+1 DebugInfo::ValueOp: One for the initial undef value (before anything
was assigned to the stack allocation), and one for each assignment
afterwards.

**Change**: Mem2ref does not always need to create this initial undef
value if it's not observable by the program. In other words, if nothing
else in the program loaded this value, at no point can an undef value be
observed by the program, so it's safe to not generate an undef unless
some other user has prompted mem2reg to create an explicit undef
already.

**Up Next**: This restores debuginfo for all but one variable in
[Internal link] The remaining case is
related to sroa splitting up debuginfo into multiple fragments, but we
don't have a way to piece them back. A separate PR will be created to
solve that case.

MODULAR_ORIG_COMMIT_REV_ID: d45a199558aaaefec21d48a4f45c83daa01fe6a8