[mojo-lang][CheckLifetimes] Remove bespoke support for variants. (#29968)
CheckLifetimes is trying to treat kgen.variants as nontrivial types,
because we put nontrivial things into them. However, everything
around this is unsafe and incorrect right now - it is a bit of a
house of cards.  We will re-evaluate this when the error
representation rework happens, but for now, CheckLifetimes should
just be "hands off", instead of differenting from the triviality
check in the Mojo parser.

This is related to Issue #28804.

MODULAR_ORIG_COMMIT_REV_ID: e88db4610af3ffb052c31cf3066a5277eda0a4a0