[KGEN] Do not crash on invalid SIMD size for splat operation (#30016)
Do not crash when you get a negative SIMD size in the splat operation,
instead give a more informative error. Note that we handle some of this
in the stdlib, but we need to have top-level struct constraints to do a
better job. Right now we handle this in the init function which is not
sufficient.

Part of [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3175178d0fa81eb50f5d633cf05b16bfb93b4636