[KGEN] Remove _setup_defines when building mojo files (#29980)
In the cmake files we were passing -D flags to the mojo files, but that
is not necessary since those flags are setup internally by KGEN itself.

This exposed a few things such as we were not passing the cmake
variables correctly to the C++ code and were getting an env which
contains

```
#kgen.env<{BUILD_TYPE = "BUILD_TYPE" : !kgen.string, KERNELS_BUILD_TYPE = "KERNELS_BUILD_TYPE" : !kgen.string, MODULAR_LLCL_MAX_PROFILING_LEVEL = 0 : index}>
```

We were also not capturing the env information into the package, which
was causing other problems.

Finally, we were overwriting the environment rather than extending it.

MODULAR_ORIG_COMMIT_REV_ID: cc362e94b026bb78530226bfcddcf5ee77cb26ba