[Support][LLCL] Move TypeID to Support. (#29737)
This patch moves the TypeID machinery to Support. Again, this is generally useful and not really LLCL-specific.

More concretely, we would like to move GenericUniquePtr over to Support, which means we need this as well.

The unit test for TypeID uses LLCL internals, so at least for the time being, we decided to leave the unit test under LLCL because things like Semaphore rightly belong inside LLCL.

This patch contains a bunch of changes to move usages that mean to use `mlir::TypeID` to spell it explicitly, since now M::TypeID is a class that behaves like-but-isn't mlir::TypeID.

MODULAR_ORIG_COMMIT_REV_ID: 3fbce4836cde6b231cb46f35f36e0a6097fdc106