[KGEN][MOGG] Move tensor util header to kgen (#30052)
To understand the parameter structure of tensors we have a C++ mirroring
helper which helps us interact with the kgen / mojo tensor. This is
needed by both graph compiler and the slicing pass of kgen, so it needs
to live in KGEN to be shared between them.

MODULAR_ORIG_COMMIT_REV_ID: 743b9aa54b149953a098da3ff03c2042a46e5e19