[Support] Fix self-assignment corruption in TensorShape (#30064)
It is valid although unusual to copy-assign something to itself.
TensorShape implicitly assumes that the value assigned is a different
object. After #30056, this results in a use-after-free. Before #30056,
the behavior was to replace the representation with uninitialized
memory, and then copy the uninitialized memory to itself. Neither of
these are correct. The easiest thing to do is to just bail out if we are
doing a self-assignment.

MODULAR_ORIG_COMMIT_REV_ID: b142a307e112bfb6bac1c01b92c82a364a92792c