[mojo-lang] Refactor parameter parsing logic a bit, NFC. (#30103)
The parameter parsing logic has grown in power and complexity,
and is spread around in different places with a bunch of parallel
state.

Introduce a new `ParsedParamSignature` type to represent all this
stuff, and start moving helpers over to it.  While here, rename
ParsedArgument.{cpp/h} to Signatures.{cpp, h} since this logic
applies to parameters as well, and also to structs.

There is more to do, but this is a step forward.

MODULAR_ORIG_COMMIT_REV_ID: c7a8f34ee280807b41079a3335e08dfd36bcbb49