[mojo-lang] Rewrite byref result emission in calls, adopt lifetimeof (#30136)
ByRef results are used by functions that return memory only types.
When calling such a function, we either way to use an existing
LValue that the call is assigning into or we want to create a new
vardecl to capture it if we can't figure that out (or it isn't safe).

This rewrites the logic for processing this to emit byref result
slots *after* all of the arguments are emitted.  This is required
for correctness when a result uses implicit lifetimes of the other
arguments, because otherwise we don't know what the type of the
result slot is supposed to be.

This rewrite also untangles logic where we had to create a temporary
byref result vardecl unconditionally, then delete it when we found
out later that we didn't want it.  We can now correctly create it
iff we actually need it.

This logic would be simpler if byref results were the last argument
instead of the first, such a change is tracked by #30134 but this
isn't blocked on it.

With all that plumbing now working correctly, we can adopt
lifetimeof in the iter method of `VariadicListMem`, whose result
type is memory only.  This dissolves another crime I had to commit
to get variadics to work, making its definition pretty reasonable!

MODULAR_ORIG_COMMIT_REV_ID: 41265582d31fed1c2b145a38e96ae2f5dbc8ec25