[KGEN] Correctly fold FMA operation (#30241)
Intermediates in FMA are evaluated in infinite precision before being
casted to ieee-float. That's not what was happening, instead we were
evaluating the intermediates using ieee-float. This was causing a
difference in results between when we would use unroll (where the
compiler has visibility in the loop and would fold) and when we do not.
Because we disable unroll with -O0, this also meant the results between
-O3 and -O0 are not the same.

This switches the folder to use APFloat::fusedMultipleAdd instead of a*b
+ c for evaluation to ensure we are doing an FMA operation.

Closes #30237
MODULAR_ORIG_COMMIT_REV_ID: 8f2d96eed8c7ba4ddc678deda236ac5b01f7b231