[Mojo] Move Parser Tests to Mojo Isolated (#29917)
This PR demonstrates how the new parser tests will operate by moving a
few of the unit tests from Mojo-parser to Mojo-isolated. Note that the
stubs have become consolidated and it will be more obvious moving
forward which dependencies we are creating between our parser tests and
the builtins because if a test lives in Mojo-isolated it cannot have a
dependency on the Kernels standard lib.

MODULAR_ORIG_COMMIT_REV_ID: a48150334af77bf8e628e8da190551079cf9019e