[mojo-stdlib] Simplify `assert_(true,false)` overloads (#30238)
Use the `Boolable` trait to avoid needing separate overloads of `assert_true`
and `assert_false`.  This requires declaring conformance of the `Boolable` trait
for `string_literal` type since its tests use `assert_true` and `assert_false`.
In turn, this introduces an ambiguous call in the `PythonObject` constructor
from `StringLiteral` type.  Fix this by delegating to the constructor taking a
`String` to avoid ambiguity with the existing `PythonObject` constructor
accepting a `Bool` since it is implicitly boolean-convertable now from a
`StringLiteral`.  In the future, we could consider reducing these constructor
sets to just work on `Stringable` types, but we'll save that for a different PR.

Note:
- As the comment in `testing.mojo` suggests, now that we have a
  `EqualityComparable` trait from the recent `Dict` work, we can apply a similar
  pattern to simplify our equality overloads in the testing module.

modular-orig-commit: 75c27f32dab362e2be091792fd3be1a1640168b5