[Support][Entitlements] Look up entitlements by name. (#30069)
In the move to name-indexed settings, we want to be able to look up Entitlements by name. This means that we need a mapping from string name to OID to be able to perform the lookup that way as well. This patch adds that capability.

MODULAR_ORIG_COMMIT_REV_ID: 9352bb580b9bf0062fba15e1566dc309765a1b62