[MOGG] Fix param deduction bug on shape / slice params (#30229)
When resolving the parameter for shapes or slices we look at the KGEN
type and pull the parameter reference from the expression used in that.
This is super super hacky and fails when the expression contains
multiple parameter references. Instead this takes a more principled
approach of attaching the parameter information onto the ToKGEN type
when we lower.

The strides are optional and we use the default parameter in a different
constructor if they are not known.

MODULAR_ORIG_COMMIT_REV_ID: cd469d3e2d252af744304943cd43af823a703226