[Telemetry] Create machineid and sessionid (redux) (#30405)
This is a resubmission of #18549.

The "machineid" is a token that applies to the local machine/container
and uniquely identifies it. It should be stable in that context.

The "sessionid" is a unique per-process token that mixes the PID with
the "machineid" to generate a unique session token.

These allow the telemetry backend to record different machines and
sessions are unique streams. Currently 99% of metrics are failing to
record, due to the presence of a duplicate resource for a given time
period. The keys will allow the backend to differentiate these cases.

MODULAR_ORIG_COMMIT_REV_ID: 8f18976f8f69cb40b7d420b62a61b87dbaec0f96