[mojo-stdlib] Rename `slice` -> `Slice` and add a `slice` function. (#30501)
Background: In Python, `slice` is a function:
https://www.w3schools.com/python/ref_func_slice.asp
it also has a type named "slice", but this is a different thing, and
certainly not our slice struct!

This patch renames the `slice` struct to `Slice` and introduces a
forwarding function named `slice`.  Our whole design in this area needs
to evolve, but this is a small step towards making things more general,
the first step of [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: f5bf233908e4316e4848cd95dc422e2b87668e5e