[Packaging] Configure and build just one target while assembling install package (#30630)
This refactor tries to simplify a bunch of things.

It gets rid of `MODULAR_CREATE_TUF_REPO` which was a workaround due to
our inability to use `MODULAR_RELEASE_PACKAGE_BUILD` with mojo-sdk due
to [Internal link] In it's place
introduce `MODULAR_STATIC_LINK_PYTHON` to signal when we're doing a
manylinux build with a statically linked Python (in which case we can
exclude mblack from mojo-sdk without consequences since that part of the
build doesn't matter for the wheels).

This allows us to configure and build one target (`install-max-tuf`)
which is now the one-stop-shop (save for the Python wheels) for creating
the install directory that eventually gets shipped to users. It depends
on downstream targets like `install-mojo-sdk` `install-max-graph-c`
`install-ai-engine-mojo` `install-ai-engine-c` and
`install-serving-tuf`. This is much better than hand-assembling the
package which was creating problems. The caveat is that Python wheels
are still outside this target because of the need to build them
separately inside a manylinux container using cibuildwheel.

MODULAR_ORIG_COMMIT_REV_ID: 0b65f578126076725d417a5cf73b8a9437f7a4e5