[Driver] Hold runtime/ctx in CompileToBinaryCtx (#30538)
Create a `CompileToBinaryCtx` and use that across drivers to ensure the
destruction order of the `Runtime` and `MLIRContext`.
Previously each driver separately ensured the declaration order of these
two, so this is a bit of code compression as well as reducing the chance
for mistakenly re-ordering them in one of the drivers.

MODULAR_ORIG_COMMIT_REV_ID: 960985407a573ce286cec0435b9ba9eb076c42d0