[Support] Add VM instance type to `system-info` (#30249)
### Changes

  - Add `getHostVMInfo()` in `MArchTarget.cpp`
  - Add VM info to `system-info` tool

### Motivation

For enhanced profile parsing & perf tracking with @maxhutch, it would be
helpful to have VM info in tracefile metadata. Adding it to
`system-info` seems like the most natural way to this, since all output
from `system-info` is included in the tracefile.

Furthermore, this is relevant system information in general. You could
have 2 machines that are otherwise identical behaving differently on AWS
& GCP due to noisy neighbor problems or other environmental issues.

### Alternative Designs

I just learned that the instance-type can also be retrieved with the
following command:
```
curl http://169.254.169.254/latest/meta-data/instance-type
```
I think this means another solution that could work would be reading
from the endpoint; this would presumably eliminate the need for
installing `cloud-utils` in the docker images. If this is a cost we are
not willing to pay I can look further into this.

### Future Work

This PR adds support for AWS VMs; we'll need to follow up with GCP
support and likely additional cloud providers in future.

### Related Issues

Closes #21245

---------

Co-authored-by: Arjun Surendran <arjun@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 49298bf317d6a10d72608e8086089d8f4a316ffd