[KGEN] Print witness tables out-of-line by default (#30699)
This makes `TypeConstantAttr`s with vtables print out-of-line by
default, BUT this adds a commandline flag to print them inline to make
testing easier. Because it is not always possible to give a good alias
name and `split-input-file` is discouraged for parser tests, this is the
only workable solution.

MODULAR_ORIG_COMMIT_REV_ID: 1e4fcd00ef91d72c5e7965b6a2e89bc56d63ac62