[mojo-vscode] Add script to check if extension matches current release (#30407)
This script will be used as part of the nightly release process to check
if we need to release a new vscode extension or not. This removes the
need to push a new release of the extension every night if it hasn't
actually changed at all.

MODULAR_ORIG_COMMIT_REV_ID: babacffd9d36a31f341ebef056685db626e503cb