[mojo-lang] Implement support for the `in` and `not in` expressions. (#30828)
This adds support for the `in` and `not in` expressions, and cleans
up some previous logic added for it. This is a dual to the famous
#30000,
giving syntax sugar to the support that patch added to the String type.

This fixes #29443

MODULAR_ORIG_COMMIT_REV_ID: 7dd9e4435b9138be26bf553ff743643332e6e9dc