[mojo-lang] Implement support for fancy slicing. (#30850)
This allows one to implement getitem operands that take arbitrary
types and treats the operands to a slice subscript as an initializer
list.  That means that given:

```
  struct S:
     fn __getitem__(self, a: YourType) -> ...
```

That a use of it like `someS[a:b:c]` will be treated as sugar for
`someS.__getitem__(YourType(a, b, c))` where missing values are
passed in as None expressions.

The implementation of this is a bit fancy: it introduces an internal
unresolved value to track the initialization and resolves it when
the contextual type is known.  In this way, it works the same way
as `{1,2,3}` initializers in C++ do, but this is only wired into
slice types.

This fixes #30257

modular-orig-commit: 79bf7f7f637038cccae7dcc6ea8b0c3b10eb9929