[Mojo][Debug] Define the options for the mojo debug command (#30693)
So far we've been using `mojo debug` as a passthrough for the regular
LLDB. However, this is not ideal for two reasons:

- This doesn't match what other commands like `mojo run` do.
- We want to merge `vscode-debug` into `mojo debug`, which requires
handling some options that LLDB simply doesn't have.

A way to address these two points is to define explicitly which options
are supported the `mojo debug` in its various variants, and still offer
an `-Xlldb` option as a backdoor for sending arguments to LLDB when in
CLI mode.

This patch doesn't implement these options yet, but I'm sending this out
already because the help message can be reviewed independently.

MODULAR_ORIG_COMMIT_REV_ID: d5ec5ac3412b09ef098b2d9fe33af6a07f57b5c6