[Mojo] Add a new __source_location magic function (#30864)
The `__source_location` magic function gives you a _SourceLocation
struct that contains the file, function, and line information where the
__source_location object was invoked from. This is going to be useful
for both enabling a more informative testing infra as well as enabling
one to add support to `assert` within the parser.

Note that the `__location` does not give you the ability to find the
location of the caller. E.g. you cannot write

```
fn assert_debug(cond:Bool, loc:_SourceLocation=__source_location(-1))
```

but that might be possible as an extension.

Closes #29373

modular-orig-commit: a41c28dd918a7b08b66e8e1c0048d1c0d2465c6f