[Mojo] [NFC] Unify Tuple Node Type Emission (#30118)
This patch resolves a crash by making it so that the code that resolves
`Tuple[T,T]` is the same code that resolve `(T,T)`. Note that I
circumvented the parsing of operands because the builtin Tuple type does
not have keyword parameter arguments so it's valid to just assume they
are all positional. Also note this does not entirely resolve
[Internal link] it just unites two
pathways so that there is no crash.

MODULAR_ORIG_COMMIT_REV_ID: 73e3438b7f1be1550bef99dc99c40ab35c8ffcc2