[Mojo] Emit error when parametric function is passed as argument (#31044)
This is never legal, so the parser should catch it. The patch also
changes an assert in the parser that could get triggered if a parametric
function is passed as an argument: if the argument is called inside a
closure, the parser would crash because closure bodies are resolved even
if the parent function had a decl resolution error. From now on, this
assert will just turn into a suppressed error if and only if the a
parent func had a resolution error.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: b1065e081f805239fb668ae014e46f86e71e7025