[docs] Fix docsite build due to JSX syntax error (#31097)
Text like `<program>` must be in code font or else docsite sees it as a
JSX (fancy HTML) tag and expects a close tag. Easy solution is to put
this sort of thing in code font. The original author clearly intended to
do so, but 4-space indent is not enough when inside a list. You must use
8 spaces.

I also took the opportunity to indent the corresponding paragraphs for
these list items so they are also nested in the HTML tags, and convert
the list to bullets, because this is not a procedure. This sometimes
required minor copy edits to keep the line lengths in bounds. Plus a
couple other edits just for clarity.

MODULAR_ORIG_COMMIT_REV_ID: a54da2fc9070623d98a693747c7f6a342fa0792d