[mojo-lang][LowerSemanticCF] Rework lowering of break/continue. (#31283)
This reworks lowering of break and continue, fixing three bugs:
1) the parser didn't diagnose break/error in a "loop else" at the
   top level of the function, we'd get a verifier error.
2) LowerSemanticCF would crash when unreachable code was after a
   break/continue in a "loop else", instead of generating a warning.
3) break/continue in a "loop else" was only handled correctly when
   immediately in the else: breaking from an if inside the loop else
   didn't work.

This shifts all responsibility of checking semantic break/continue to
the LowerSemanticCF pass, so the parser isn't responsible for #1.
This simplifies the verifier and the parser and avoids duplication
of concerns between the parser/verifier/LowerSemanticCF.

This then goes on to modernize LowerSemanticCF a bit - it now uses a
struct instead of just recursive functions.  This is just a cleanup
because the code has gotten bigger.

The final more significant change is a rewrite of the lowering of loops
in the lowerLITLoop function.  Now it uses the "splice" operation to
move operations across blocks en-masse, and uses the existing recursive
traversal to rewrite the break/continue ops.

MODULAR_ORIG_COMMIT_REV_ID: c5a9b40808fdb48c0f74a901ab4a1f9f27ff8f54