[Mojo] Improve error message when returning type instead of instance (#31297)
This is a very common type/mistake, but our error message was extremely
poor: `cannot implicitly convert 'Foo' value to 'Foo' in return value`.
This patch improves that by explicitly checking if the returned value is
a type value that is equivalent to the expected type, and emits a more
meaningful error message along with a hint on how to fix it.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 697578b69dff70114f8c1df040483a434cb45a87