[Support][Telemetry] Telemetry needs Settings. (#30482)
This patch adopts the Settings structure for creating a TelemetryContext. It also adopts the pattern that the Settings structure started for adding an instance of yourself to the Context structure.

This patch contains a *lot* of code duplication, which I'm not thrilled about, but it's in the pursuit of deleting a lot of the original code in the future.

MODULAR_ORIG_COMMIT_REV_ID: f4bb8031ee030cb2a4e40482da57150b94abf754